requirements.txt
Flask
Flask-SQLAlchemy
boto3
pycryptodome
gunicorn